backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
python-multipart
PyPDF2
